Iterate over each directory in the file structure and slight refactor
Previously, we were only running the 0.13upgrade tool over the root
directory of each repository. We actually need it to run on every
terraform file. This change includes walking through the directory
structure and placing each directory in a slice, then iterating the
binary over each directory.

This does introduce some added complexity and therefore increases the
time it takes for the script to complete.